tests/wpt/tests/soft-navigation-heuristics/supported-entry-types.tentative.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<title>Soft navigations are a supported entry type</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  promise_test(async () => {
    assert_true(PerformanceObserver.supportedEntryTypes.includes(
      "soft-navigation"));
  }, "Soft navigations are a supported entry type");
</script>
